Create notification model
